requirements.txt
discord.py>=2.3
aiohttp>=3.9
python-dotenv
beautifulsoup4
lxml
